requirements.txt
fastapi
uvicorn
pyrogram
python-dotenv
